separate vw and pytorch
