+++ src/learn_to_pick/__init__.py
@@ -5,12 +5,9 @@ from learn_to_pick.base import (
     BasedOn,
     Embed,
     Featurizer,
-    ModelRepository,
     Policy,
     SelectionScorer,
     ToSelectFrom,
-    VwPolicy,
-    VwLogger,
     embed,
 )
 from learn_to_pick.pick_best import (
@@ -21,9 +18,13 @@ from learn_to_pick.pick_best import (
     PickBestSelected,
 )
 
-from learn_to_pick.byom.pytorch_policy import PyTorchPolicy
 
-from learn_to_pick.byom.pytorch_feature_embedder import PyTorchFeatureEmbedder
+from learn_to_pick.vw.policy import VwPolicy
+from learn_to_pick.vw.model_repository import ModelRepository
+from learn_to_pick.vw.logger import VwLogger
+
+from learn_to_pick.pytorch.policy import PyTorchPolicy
+from learn_to_pick.pytorch.pytorch_feature_embedder import PyTorchFeatureEmbedder
 
 
 def configure_logger() -> None:
@@ -52,11 +53,11 @@ __all__ = [
     "SelectionScorer",
     "AutoSelectionScorer",
     "Featurizer",
-    "ModelRepository",
     "Policy",
     "PyTorchPolicy",
     "PyTorchFeatureEmbedder",
-    "VwPolicy",
-    "VwLogger",
     "embed",
+    "ModelRepository",
+    "VwPolicy",
+    "VwLogger"
 ]
